requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
openai==0.27.0
python-dotenv==1.0.0
Werkzeug==2.2.3
